+++ CHANGES.rst
@@ -7,6 +7,8 @@ Unreleased
 
 - Make the ``config get`` commands respect the output format option.
 
+- Fixed API redirect and error response bug for ``monitoring grafana`` command.
+
 0.11.0 - 2019/04/17
 ===================